requirements.txt
Django==2.1.15
django-ckeditor==6.0.0
django-js-asset==1.2.2
PyMySQL==0.9.3
Pillow==8.1.0
pytz==2021.1
requests==2.22.0